# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "BaseResponse_CourseResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CourseResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message"
        ],
        "title": "BaseResponse[CourseResponse]",
        "type": "object"
      },
...
      "BaseResponse_list_CourseResponse__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CourseResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message"
        ],
        "title": "BaseResponse[list[CourseResponse]]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Lumina AI Study Platform Backend API",
    "title": "Lumina API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/courses/": {
      "get": {
        "description": "Lists the active courses the caller is allowed to read.",
        "operationId": "get_courses_api_courses__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_list_CourseResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Courses",
        "tags": [
          "Courses"
        ]
      },
      "post": {
        "description": "Creates a course owned by the authenticated user.\n\nOwnership comes from the verified token, never from the request body, so a\ncaller cannot create a course under another account.",
        "operationId": "create_course_api_courses__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_CourseResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Course",
        "tags": [
          "Courses"
        ]
      }
    },
...
    "/api/courses/{course_id}/exam-mode/analysis": {
      "get": {
        "operationId": "get_exam_analysis_api_courses__course_id__exam_mode_analysis_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "output_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Output Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_ExamAnalysisView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Course or exam topic analysis not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Exam Analysis",
        "tags": [
          "Exam Mode"
        ]
      },
      "post": {
        "operationId": "analyse_exam_sources_api_courses__course_id__exam_mode_analysis_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_ExamAnalysisResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "No processed course material is available"
          },
          "401": {
            "description": "Authentication required"
          },
          "402": {
            "description": "Insufficient credits"
          },
          "404": {
            "description": "Course or document not found"
          },
          "409": {
            "description": "No course material matched the request, the course material is not searchable yet, or a selected document is still being processed"
          },
          "422": {
            "description": "Invalid exam analysis request"
          },
          "429": {
            "description": "AI provider or per-user generation rate limited"
          },
          "503": {
            "description": "AI provider or course search unreachable"
          },
          "504": {
            "description": "AI provider timed out"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Analyse Exam Sources",
        "tags": [
          "Exam Mode"
        ]
      }
    },
    "/api/courses/{course_id}/exam-mode/analysis/rescan": {
      "post": {
        "description": "Analyse the selected sources again, leaving every existing plan alone.\n\nA rescan is its own route rather than a flag because it is priced\ndifferently, and a price a client must not hardcode has to be readable from\nthe credit policy by its own name. It also keeps its own rate-limit bucket,\nso a cheap rescan cannot drain the expensive first analysis.",
        "operationId": "rescan_exam_sources_api_courses__course_id__exam_mode_analysis_rescan_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_ExamAnalysisResult_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "No processed course material is available"
          },
          "401": {
            "description": "Authentication required"
          },
          "402": {
            "description": "Insufficient credits"
          },
          "404": {
            "description": "Course or document not found"
          },
          "409": {
            "description": "No course material matched the request, the course material is not searchable yet, or a selected document is still being processed"
          },
          "422": {
            "description": "Invalid exam analysis request"
          },
          "429": {
            "description": "AI provider or per-user generation rate limited"
          },
          "503": {
            "description": "AI provider or course search unreachable"
          },
          "504": {
            "description": "AI provider timed out"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Rescan Exam Sources",
        "tags": [
          "Exam Mode"
        ]
      }
    },
    "/api/courses/{course_id}/exam-mode/analysis/{output_id}/questions": {
      "get": {
        "operationId": "list_past_exam_questions_api_courses__course_id__exam_mode_analysis__output_id__questions_get",
        "parameters": [
          {
            "in": "path",
            "name": "output_id",
            "required": true,
            "schema": {
              "title": "Output Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "topic_key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic Key"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_ExamQuestionPage_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Course or exam topic analysis not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Past Exam Questions",
        "tags": [
          "Exam Mode"
        ]
      }
    },
...
    "/api/courses/{course_id}/exam-mode/plans/{output_id}": {
      "get": {
        "description": "Return one stored plan exactly as it was written.\n\nA database read and nothing else: no provider, no retrieval, no embedding,\nno charge, and no write. A plan whose exam date has long passed still\nopens, because it is a study resource rather than a countdown.",
        "operationId": "get_exam_plan_api_courses__course_id__exam_mode_plans__output_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "output_id",
            "required": true,
            "schema": {
              "title": "Output Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_ExamPlanView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Course or exam plan not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Exam Plan",
        "tags": [
          "Exam Mode"
        ]
      }
    },
    "/api/courses/{course_id}/exam-mode/sources": {
      "get": {
        "operationId": "list_exam_sources_api_courses__course_id__exam_mode_sources_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "title": "Course Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse_ExamSourceInventory_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Course not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Exam Sources",
        "tags": [
          "Exam Mode"
        ]
      }
    },
...
  }
}